# Conflicts:
#	src/app/reducers/templates/redux/template.effects.ts
#	src/app/reducers/vm/redux/vm-creation.effects.ts
#	src/app/template/containers/template-filter-selector.container.ts
#	src/app/template/shared/iso.service.ts
#	src/app/template/template-sidebar/template-description/template-description.component.html
#	src/app/template/template/card-item/template-card-item.component.html
#	src/app/template/template/row-item/template-row-item.component.html
#	src/app/vm/vm-creation/vm-creation.component.ts
#	src/testutils/mocks/model-services/entity-data.spec.ts

+++ src/app/vm/vm-sidebar/storage-detail/volumes/snapshot/snapshot-modal.container.ts
@@ -1,28 +1,27 @@
-import {
-  Component,
-  Inject,
-  OnInit
-} from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { State } from '../../../../../reducers/index';
-import {
-  MAT_DIALOG_DATA,
-  MatDialogRef
-} from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 import { Volume } from '../../../../../shared/models/volume.model';
 import { Snapshot } from '../../../../../shared/models/snapshot.model';
+// tslint:disable-next-line
+import { SnapshotActionService } from '../../../../../snapshot/snapshots-page/snapshot-list-item/snapshot-actions/snapshot-action.service';
 import { WithUnsubscribe } from '../../../../../utils/mixins/with-unsubscribe';
 
 import * as volumeActions from '../../../../../reducers/volumes/redux/volumes.actions';
 import * as snapshotActions from '../../../../../reducers/snapshots/redux/snapshot.actions';
 import * as fromVolumes from '../../../../../reducers/volumes/redux/volumes.reducers';
+import { DialogService } from '../../../../../dialog/dialog-service/dialog.service';
 
 @Component({
   selector: 'cs-snapshot-modal-container',
   template: `
     <cs-snapshot-modal
       [volume]="volume$ | async"
-      (onSnapshotDelete)="snapshotDeleted($event)"
+      (onTemplateCreate)="onTemplateCreate($event)"
+      (onVolumeCreate)="onVolumeCreate($event)"
+      (onSnapshotRevert)="onSnapshotRevert($event)"
+      (onSnapshotDelete)="onSnapshotDelete($event)"
     >
     </cs-snapshot-modal>`,
 })
@@ -35,6 +34,8 @@ export class SnapshotModalContainerComponent extends WithUnsubscribe() implement
     @Inject(MAT_DIALOG_DATA) data,
     public dialogRef: MatDialogRef<SnapshotModalContainerComponent>,
     private store: Store<State>,
+    private dialogService: DialogService,
+    private snapshotActionService: SnapshotActionService
   ) {
     super();
     this.store.dispatch(new volumeActions.LoadSelectedVolume(data.volumeId));
@@ -53,7 +54,24 @@ export class SnapshotModalContainerComponent extends WithUnsubscribe() implement
       });
   }
 
-  public snapshotDeleted(snapshot: Snapshot) {
-    this.store.dispatch(new snapshotActions.DeleteSnapshot(snapshot));
+  public onTemplateCreate(snapshot: Snapshot) {
+    this.snapshotActionService.showTemplateCreationDialog(snapshot);
+  }
+
+  public onVolumeCreate(snapshot: Snapshot) {
+    this.snapshotActionService.showVolumeCreationDialog(snapshot);
+  }
+
+  public onSnapshotDelete(snapshot: Snapshot): void {
+    this.dialogService.confirm({ message: 'DIALOG_MESSAGES.SNAPSHOT.CONFIRM_DELETION' })
+      .onErrorResumeNext()
+      .filter(res => Boolean(res))
+      .subscribe(() => {
+        this.store.dispatch(new snapshotActions.DeleteSnapshot(snapshot));
+      });
+  }
+
+  public onSnapshotRevert(snapshot: Snapshot): void {
+    this.store.dispatch(new snapshotActions.RevertVolumeToSnapshot(snapshot));
   }
 }
